requirements.txt
fastapi 
uvicorn
reaction
scipy==1.0.0
matplotlib
Pillow==5.3.0
opencv-python
scikit-image
torch==1.4.0
torchvision==0.5.0
cython
